FIxes typo in print statement for day 5
The day 5 python executable file contains helpful print statements at the bottom that communicate the answer to part 1 and part 2 of the puzzle. Currently, the second print statement, meant to print the answer for part 2, is incorrectly labeled as the answer for part 1. This commit aims to remedy this pressing issue.